requirements.txt
torch
torchvision
timm
mlflow
optuna
numpy
pandas
matplotlib
Pillow